Attendance_system/requirements.txt
flask
torch
keras
mtcnn
pillow
numpy


